+++ CHANGELOG.md
@@ -5,8 +5,9 @@
 * `\Tools\Exceptionist::methodExists()` throws the new `Tools\Exception\MethodNotExistsException`;
 * by default, the `\Tools\Exceptionist::____callStatic()` magic method now throws an `ErrorException`;
 * improved the description of many `Exceptionist` magic methods;
-* `\Tools\TestSuite\TestTrait::assertException()` correctly ignores the  deprecations in determining 
-  the exception;
+* `\Tools\TestSuite\TestTrait::assertException()` new performs a strict comparison and does not
+  consider parent classes (`ErrorException` != `Exception`). It correctly ignores the  deprecations
+  in determining the exception;
 * added many `@template` tags to the `Exceptionist`, that now always refers to `Exception` and no
   longer to `Throwable`;
 * improved the `array_clean()` global function.